+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [0.14.0] - 2018-04-13
+
 ### Added
 
 - Prompt user to create billing profile if missing during resource create